add: mejoras en permisos

+++ app/Http/Controllers/PerfilController.php
@@ -11,17 +11,26 @@ class PerfilController extends Controller
 {
     public function index()
     {
+        if (!checkPermiso('perfiles.is_read')) {
+           return redirect()->back()->with('error', 'No tienes permisos para realizar esta acción');
+        }
         $perfiles = Perfil::all();
         return view('pages.perfiles.index', compact('perfiles'));
     }
 
     public function create()
     {
+        if (!checkPermiso('perfiles.is_create')) {
+           return redirect()->back()->with('error', 'No tienes permisos para realizar esta acción');
+        }
         return view('pages.perfiles.create');
     }
 
     public function store(Request $request)
     {
+        if (!checkPermiso('perfiles.is_create')) {
+           return redirect()->back()->with('error', 'No tienes permisos para realizar esta acción');
+        }
         $request->validate([
             'nombre' => 'required|unique:perfiles,nombre',
             'descripcion' => 'nullable|string',
@@ -38,12 +47,18 @@ class PerfilController extends Controller
 
     public function edit(Perfil $perfil)
     {
+        if (!checkPermiso('perfiles.is_update')) {
+           return redirect()->back()->with('error', 'No tienes permisos para realizar esta acción');
+        }
         $permisos = Permiso::all(); // <-- asegúrate de importar el modelo
         return view('pages.perfiles.edit', compact('perfil', 'permisos'));
     }
 
     public function update(Request $request, Perfil $perfil)
     {
+        if (!checkPermiso('perfiles.is_update')) {
+           return redirect()->back()->with('error', 'No tienes permisos para realizar esta acción');
+        }
         $request->merge([
             'nombre' => strtolower(trim($request->nombre))
         ]);
@@ -74,6 +89,9 @@ class PerfilController extends Controller
 
     public function destroy(Perfil $perfil)
     {
+        if (!checkPermiso('perfiles.is_delete')) {
+           return redirect()->back()->with('error', 'No tienes permisos para realizar esta acción');
+        }
         $perfil->delete();
         return redirect()->route('perfiles.index')->with('success', 'Perfil eliminado.');
     }
